% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{akaike.wts}
\alias{akaike.wts}
\title{Compute Akaike weights from AIC scores}
\usage{
akaike.wts(aa)
}
\arguments{
\item{aa}{vector of AIC or AICc scores}
}
\value{
a vector of Akaike weights
}
\description{
Compute Akaike weights from AIC scores
}
\details{
This function converts a vector of AIC or AICc scores to Akaike
weights, a vector that summarizes proportional model support.
}
\examples{
akaike.wts(c(10, 14, 20))

}
\seealso{
\code{\link{compareModels}}
}
