% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{as.paleoTSfit}
\alias{as.paleoTSfit}
\title{Create a \code{paleoTSfit} object}
\usage{
as.paleoTSfit(logL, parameters, modelName, method, K, n, se)
}
\arguments{
\item{logL}{model log-likelihood}

\item{parameters}{model parameter estimates}

\item{modelName}{model name}

\item{method}{parameterization, either "AD" or "Joint"}

\item{K}{number of model parameters}

\item{n}{sample size}

\item{se}{standard errors of parameter estimates}
}
\value{
a \code{paleoTSfit} object
}
\description{
Create a \code{paleoTSfit} object
}
\note{
All model-fitting functions use this function to package the resulting
  data-model fits. Users will not need to call this function.
}
