% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cantius_L}
\alias{cantius_L}
\title{Time-series of the length of lower first molar for the Cantius lineage}
\format{a \code{paleoTS} object with the data}
\source{
Clyde, W. C. and P. D. Gingerich (1994). Rates of evolution
in the dentition of early Eocene Cantius: comparison of size and shape.
\emph{Paleobiology} 20(4): 506-522.
}
\usage{
cantius_L
}
\description{
Time-series of the length of lower first molar for the Cantius lineage
}
\keyword{datasets}
