% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dorsal.spines}
\alias{dorsal.spines}
\title{Time-series of dorsal spine data from a fossil stickleback lineage}
\format{
a \code{paleoTS} object of the mean number of dorsal spines (log-transformed)
}
\source{
Bell, M.A., M.P. Travis and D.M. Blouw 2006. Inferring natural
selection in a fossil threespine stickleback. \emph{Paleobiology} \bold{ 32}:562-577. \cr \cr
Hunt, G., M. A. Bell and M. P. Travis (2008). Evolution toward a new adaptive optimum:
phenotypic evolution in a fossil stickleback lineage. \emph{Evolution} 62(3): 700-710.
}
\usage{
dorsal.spines
}
\description{
Time-series of dorsal spine data from a fossil stickleback lineage
}
\keyword{datasets}
