% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{IC}
\alias{IC}
\title{Compute Information Criteria}
\usage{
IC(logL, K, n = NULL, method = c("AICc", "AIC", "BIC"))
}
\arguments{
\item{logL}{log-likelihood}

\item{K}{number of parameters}

\item{n}{sample size}

\item{method}{either "AIC", "AICc", or "BIC"}
}
\value{
the value of the specified information criterion
}
\description{
Compute Information Criteria
}
\note{
This function is used internally by the model-fitting functions. It
  will not generally be called directly by the user.
}
