% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleoTS-package.R
\docType{package}
\name{paleoTS-package}
\alias{paleoTS}
\alias{paleoTS-package}
\title{paleoTS: Analyze Paleontological Time-Series}
\description{
Facilitates analysis of paleontological sequences of trait values. Functions are provided to fit, using maximum likelihood, simple evolutionary models (including unbiased random walks, directional evolution,stasis, Ornstein-Uhlenbeck, covariate-tracking) and complex models (punctuation, mode shifts).
}
\author{
\strong{Maintainer}: Gene Hunt \email{hunte@si.edu} (\href{https://orcid.org/0000-0001-6430-5020}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item John Fricks \email{jfricks@asu.edu} [contributor]
}

}
\keyword{internal}
