% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{print.paleoTSfit}
\alias{print.paleoTSfit}
\title{Print a paleoTSfit object}
\usage{
\method{print}{paleoTSfit}(x, ...)
}
\arguments{
\item{x}{a paleoTSfit object}

\item{...}{other arguments for other print methods}
}
\value{
None; this function is used only to print
}
\description{
Print a paleoTSfit object
}
\examples{
y <- sim.punc(theta = c(0, 2), omega = c(0.1, 0.1))
wg <- fitGpunc(y)
print(wg)
}
